feat: Support Mixtral-8x7B
